+++ src/main/java/jscover/json/JSONDataSaver.java
@@ -345,7 +345,6 @@ package jscover.json;
 import jscover.util.IoUtils;
 
 import java.io.File;
-import java.io.StringReader;
 import java.util.List;
 import java.util.SortedMap;
 import java.util.TreeMap;
@@ -360,13 +359,13 @@ public class JSONDataSaver {
         if (jsonFile.exists()) {
             String existingJSON = IoUtils.toString(jsonFile);
             extraData.putAll(jsonDataMerger.mergeJSONCoverageData(existingJSON, data));
-            IoUtils.copy(new StringReader(jsonDataMerger.toJSON(extraData)), jsonFile);
+            IoUtils.copy(jsonDataMerger.toJSON(extraData), jsonFile);
         } else if (unloadJSData != null) {
             //Only scan for unloaded JS if JSON not saved before
             extraData.putAll(jsonDataMerger.createEmptyJSON(unloadJSData));
             extraData.putAll(jsonDataMerger.jsonToMap(data));
-            IoUtils.copy(new StringReader(jsonDataMerger.toJSON(extraData)), jsonFile);
+            IoUtils.copy(jsonDataMerger.toJSON(extraData), jsonFile);
         } else
-            IoUtils.copy(new StringReader(data), jsonFile);
+            IoUtils.copy(data, jsonFile);
     }
 }